Changed location of component and added attempted to add recaptcha
