+++ public/_document.js
@@ -21,6 +21,11 @@ export default function Document() {
           src="https://cdn.jsdelivr.net/npm/@emailjs/browser@4/dist/email.min.js"
         ></script>
         <script src="https://www.google.com/recaptcha/api.js" async defer></script> 
+        <script
+          src={`https://www.google.com/recaptcha/api.js?render=${process.env.NEXT_PUBLIC_RECAPTCHA_KEY}`}
+          async
+          defer
+        ></script>
       </Head>
       <body className="antialiased">
         <Main />
